requirements.txt
## The following requirements were added by pip freeze:
flask
python-socketio
requests
base58
solders
python-dotenv